BigInteger and BigDecimal now long and double in java code, but BigInteger and double in DB. changeset now divided in 3 files and translated in ".xml".

+++ src/main/java/com/example/employee/dto/JsonEmployee.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import java.math.BigInteger;
 import java.time.LocalDate;
 
-public record JsonEmployee(BigInteger id,
+public record JsonEmployee(Long id,
                            @NotNull String firstName,
                            @NotNull String lastName,
                            @NotNull String position,
